Audit 12b: drobne chyby chovani z uklidu
Zamereni po rozbiti, liny worker sveta a uvolneni po Save and Quit,
casovace polozek skutecnym casem, kurzor v kontejneru navrch, undo bez
prazdnych kroku, nahled stromu bez terenu, koruna ne pod zem, fit
pulenim, baleni animace a yaw, EXTENSIBLE WAV podle subformatu, F3 s
Locale.ROOT, pole Create ukazuji konec textu, trackpad, a testy, ktere
driv nic nehlidaly nebo nezastavovaly svety.

+++ src/main/java/mc/BiomeTuning.java
@@ -499,9 +499,9 @@ public final class BiomeTuning {
             throw new IllegalArgumentException("koren neni objekt");
         }
 
-        if(root.get("format") instanceof Double format && format.intValue() > FORMAT)
+        if(root.get("format") instanceof Double format && format > FORMAT)
         {
-            problems.add("format " + format.intValue() + " je novejsi nez " + FORMAT
+            problems.add("format " + format + " je novejsi nez " + FORMAT
                     + " - co nezname, se preskoci");
         }
